requirements.txt
aiofiles==23.2.1
aiogram==3.4.1
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
frozenlist==1.4.1
idna==3.6
magic-filter==1.0.12
multidict==6.0.5
numpy==1.26.4
pandas==2.2.1
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic_core==2.14.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
typing_extensions==4.10.0
tzdata==2024.1
yarl==1.9.4
